Renamed the library main package
