+++ LibraryRateMe/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="com.androidsx.libraryrateme.DialogRateMe"
+            android:name="com.androidsx.rateme.DialogRateMe"
             android:label="@string/title_activity_dialog" >
         </activity>
     </application>
